Rename sweetgreen file to sweetgreenGeoJson and fix import in App
